requirements.txt
werkzeug==3.0.1
scapy==2.5.0
streamlit==1.31.0
requests==2.31.0
pyyaml==6.0.1
rich==13.7.0
gunicorn==21.2.0
pandas>=2.0.0
flask>=2.3.0
PyYAML>=6.0
